pyproject.toml
[project]
name = "Narr_ai_tive"
version = "0.1.0"
description = "Novel Content Embeddings & Story Generator"
requires-python = ">=3.8"
dependencies = [
    "google-generativeai>=0.3.0",
    "pathlib",
    "tqdm",
    "numpy",
    "ratelimit",
    "pyyaml",
    "requests",
    "scikit-learn",
    "nltk>=3.6.0",
    "rouge>=1.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "rich>=13.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
